+++ enterpriseusagereportsv1/enterprise_usage_reports_v1_examples_test.go
@@ -134,7 +134,7 @@ var _ = Describe(`EnterpriseUsageReportsV1 Examples Tests`, func() {
 				panic(err)
 			}
 			b, _ := json.MarshalIndent(reports, "", "  ")
-			fmt.Println(string(b))
+			fmt.Printf("\nGetResourceUsageReport() result:\n%s\n", string(b))
 
 			// end-get_resource_usage_report