+++ resourcemanagerv2/resource_manager_v2_suite_test.go
@@ -17,9 +17,10 @@
 package resourcemanagerv2_test
 
 import (
+	"testing"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"testing"
 )
 
 func TestResourceManagerV2(t *testing.T) {