+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [0.17.20](https://github.com/IBM/platform-services-go-sdk/compare/v0.17.19...v0.17.20) (2021-03-25)
+
+
+### Bug Fixes
+
+* **Global Search:** update service after recent API changes  ([#99](https://github.com/IBM/platform-services-go-sdk/issues/99)) ([cafd943](https://github.com/IBM/platform-services-go-sdk/commit/cafd943f7e5cf7ff6db47ad49546f0f49f5480a3))
+
+## [0.17.19](https://github.com/IBM/platform-services-go-sdk/compare/v0.17.18...v0.17.19) (2021-03-25)
+
+
+### Bug Fixes
+
+* add labels to examples ([#97](https://github.com/IBM/platform-services-go-sdk/issues/97)) ([35e8342](https://github.com/IBM/platform-services-go-sdk/commit/35e8342ffe1d5a18ffee3d36bd495dd49d1f9b07))
+
+## [0.17.18](https://github.com/IBM/platform-services-go-sdk/compare/v0.17.17...v0.17.18) (2021-03-18)
+
+
+### Bug Fixes
+
+* **Resource Manager:** update service after recent API changes and generate examples ([#92](https://github.com/IBM/platform-services-go-sdk/issues/92)) ([a049b18](https://github.com/IBM/platform-services-go-sdk/commit/a049b184d4cdc280497f6460680abdc83dea694d))
+
+## [0.17.17](https://github.com/IBM/platform-services-go-sdk/compare/v0.17.16...v0.17.17) (2021-03-11)
+
+
+### Bug Fixes
+
+* **IAM Identity:** update service after recent API changes ([#93](https://github.com/IBM/platform-services-go-sdk/issues/93)) ([c56e295](https://github.com/IBM/platform-services-go-sdk/commit/c56e295e94a715ae0bcd33d1ae2115fe368616b9))
+
 ## [0.17.16](https://github.com/IBM/platform-services-go-sdk/compare/v0.17.15...v0.17.16) (2021-03-05)